Better router system.
There is now a json config file that defines the routes and allows to pass variables in a beautiful manner.

There is also a way to generate url's with the method url() of the new router.

The new router system is not well commented and it should be done soon.